Fixed bug grabbing "real" title from Mangadex API
Copied from Issue Comment: "what was causing this issue was me hardcoding the function that grabs the "real" title from a search string to be labeled as "en", rather than just grabbing the one that was available. Because Nagatoro had its title labeled as "jp" it was unable to find the "real" title. This is why it would still return information about the manga but prevented you from tracking it. "